Capsim Directory With Topologies for QPSK Communication Link Block Diagrams, TCL Scriots
Directory Capsim contains the directory Topologiies.
The Topologies directory contains the block diagrams for modeling and simulating the QPSK Digital Communications Links. It also contains the TCL script for BER versus SNR calculation and tabulation.
The Capsim directory also has the Makefile for building 'capsim' which is the executable for simulation.